Improve generics.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@789045 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/widget/src/org/ofbiz/widget/form/ModelFormField.java
@@ -1473,7 +1473,7 @@ public class ModelFormField {
         }
 
         public static int findFieldTypeFromName(String name) {
-            Integer fieldTypeInt = (Integer) FieldInfo.fieldTypeByName.get(name);
+            Integer fieldTypeInt = FieldInfo.fieldTypeByName.get(name);
             if (fieldTypeInt != null) {
                 return fieldTypeInt.intValue();
             } else {
